Class OpsDescription contains a list of modules to know which module is already registered at this OPS.
